rename entities
