+++ src/main/java/com/spring/universidad/cryptop2p/modelo/entities/Transaction.java
@@ -1,8 +1,8 @@
-package com.spring.universidad.cryptop2p.modelo.entidades;
+package com.spring.universidad.cryptop2p.modelo.entities;
 
 
 
-import com.spring.universidad.cryptop2p.modelo.entidades.numeradores.CryptoEnum;
+import com.spring.universidad.cryptop2p.modelo.entities.numeradores.CryptoEnum;
 
 import javax.persistence.*;
 import java.io.Serializable;
