+++ src/main/java/com/spring/universidad/cryptop2p/modelo/entities/User.java
@@ -1,4 +1,4 @@
-package com.spring.universidad.cryptop2p.modelo.entidades;
+package com.spring.universidad.cryptop2p.modelo.entities;
 
 import javax.persistence.*;
