+++ src/main/java/com/spring/universidad/cryptop2p/modelo/entities/controller/UserController.java
@@ -1,10 +1,10 @@
-package com.spring.universidad.cryptop2p.modelo.entidades.controller;
+package com.spring.universidad.cryptop2p.modelo.entities.controller;
 
 import com.spring.universidad.cryptop2p.services.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import com.spring.universidad.cryptop2p.modelo.entidades.dto.UserRegisterDto;
+import com.spring.universidad.cryptop2p.modelo.entities.dto.UserRegisterDto;
 import javax.validation.Valid;
 
 @RestController
