+++ src/main/java/com/spring/universidad/cryptop2p/modelo/entities/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package com.spring.universidad.cryptop2p.modelo.entidades.repository;
+package com.spring.universidad.cryptop2p.modelo.entities.repository;
 
-import com.spring.universidad.cryptop2p.modelo.entidades.User;
+import com.spring.universidad.cryptop2p.modelo.entities.User;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
